% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my1func.R
\name{buscar_marmitas_veganas}
\alias{buscar_marmitas_veganas}
\title{Encontra marmitas veganas}
\usage{
buscar_marmitas_veganas()
}
\value{
Uma tibble com todas as marmitas disponiveis
}
\description{
Encontra marmitas veganas
}
\examples{
buscar_marmitas_veganas()
}
