% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my1func.R
\name{encontrar_pior_ano_time}
\alias{encontrar_pior_ano_time}
\title{Encontra pior ano do Brasilirao de um time}
\usage{
encontrar_pior_ano_time(time = "Cruzeiro")
}
\arguments{
\item{time}{Uma string, o nome de um time ('Cruzeiro' por padrao)}
}
\value{
Uma tibble com tres colunas
}
\description{
Usando a base do Brasileirao, encontra o ano com o menor
numero de vitorias desse time
}
\examples{
encontrar_pior_ano_time(time = "Cruzeiro")
}
