% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my1func.R
\name{melhor_time_temporada}
\alias{melhor_time_temporada}
\title{Encontra o melhor time da temporada}
\usage{
melhor_time_temporada(ano = NULL)
}
\arguments{
\item{ano}{Uma string, um valor de ano. \code{Default = NULL} retorna todos os anos}
}
\value{
Uma tibble com tres colunas
}
\description{
Encontra o melhor time da temporada
}
\examples{
melhor_time_temporada(ano = '2003')
}
