% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{partidas_brasileirao}
\alias{partidas_brasileirao}
\title{Titulo da base}
\format{
Uma lista que descreve as colunas:
\describe{
\item{col1}{Descricao da coluna 1}
\item{col2}{Descricao da coluna 2}
...
}
}
\source{
Origem dos dados
}
\usage{
partidas_brasileirao
}
\description{
Descricao da base
}
\keyword{datasets}
